:bug: Fix status check for pre-approved travel request submission documents.
Previously I was checking the "submission" status, but I really needed to be checking the "pre-approval" status.
Why? Because pre-approval requests can be individually rejected, with the overal submission still being approved.
